[WIP] Issue #25

+++ database/schemas.go
@@ -6,8 +6,8 @@ type Torrent struct {
 	InfoHash   string `gorm:"varchar(32), not null"`
 	Name       string `gorm:"not null"`
 	Downloaded int    `gorm:"not null"`
-	Seeders    int    `gorm:"not null"`
-	Leechers   int    `gorm:"not null"`
+	Seeders    int64  `gorm:"not null"`
+	Leechers   int64  `gorm:"not null"`
 	AddedBy    string `gorm:"varchar(15)"`
 	DateAdded  int64
 }
